lab/requirements.txt
opencv_contrib_python==4.5.2.54
Pillow==8.3.1
pyzmq==22.1.0
